requirements.txt
openai==0.28.0
pdf2image==1.17.0
Pillow==11.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
python_docx==1.1.2
streamlit==1.41.1
transformers==4.48.2
